Wave 40: 3513 -> 3563 (+50), bytes 49.22% -> 50.12% -- past half the ROM
Ten agents (cap 3 concurrent, 5-6 functions each), address-adjacent blocks
0x08003-0x0801E: 50 of 64 attempted verified by exit status, promoted into
35 units, both builds reproduce 14dd0b22.

What was learned:
- Exemplar-dense small-function blocks keep converting at 100% (five
  batches went perfect, incl. every extension trio taken); the 14 misses
  all sit in 150-250B functions needing fresh struct/function-pointer
  typing. Every miss left best.c + score for the next pre-flight.
- sub_0801DAB0's wave-35 empty-parameter-list correction reaffirmed: a new
  (void) definition lost to the caller-side fact (c_08039BDC passes a real
  ProcPtr). proto_check caught it pre-build; definition edited, re-matched.
- sub_0801DC04 returns s16 (its body narrows at #0x10), so promoted caller
  c_08015438's s8 truncation is the caller's own cast -- rewritten
  explicitly, byte-neutral, unit re-verified.
- A mid-wave redefinition between two agents sharing block 0x0801B was a
  REAL duplicate again, not a torn read -- second wave running to the
  wave-33 lesson; the re-run-once heuristic gave the right recovery.
- Notable near-misses for wave 41: proc-runner twins sub_080149C0 /
  sub_08014A5C at 97.4/97.7% (one shared shape, likely one shared fix);
  sub_08019A60 diverges at +0x4 (parameter/context model, not body).

+++ src/decomp/c_08015438.c
@@ -14,10 +14,16 @@
  * The two result locals are DIFFERENT types and both readouts are in the
  * epilogue. `i` is `s8`: the ROM keeps the zero-extended copy in sl and
  * re-signs it at the return (`lsls #0x18; asrs #0x18`), which is PROMOTE_MODE
- * on an s8 local. `j` is `int`: it gets ONE `lsls #0x18; asrs #0x18` -- the
- * call-site re-narrowing of sub_0801DC04's s8 return into a wide local -- and
+ * on an s8 local. `j` is `int`: it gets ONE `lsls #0x18; asrs #0x18`, and
  * every later use, including `return j`, reads that register with no further
  * extension. Declaring `j` as `s8` adds a second pair.
+ *
+ * Wave 40, W40-I: that single pair used to be explained as agbcc re-narrowing
+ * an `s8` return from sub_0801DC04, and it is now written as an explicit
+ * `(s8)` cast instead. sub_0801DC04's own body (matched this wave) narrows the
+ * value it returns at #0x10, not #0x18, so it cannot be an s8-returning
+ * function; the truncation is this caller's. The emitted bytes here are
+ * unchanged either way -- re-verified with trymatch.
  *
  * Parameter 4 arrives as `void *` from the already-promoted sub_08015410 and
  * its only use here is `lsls #0x10; asrs #0x10`, so the cast chain is written
@@ -35,7 +41,7 @@ s8 sub_08015438(void *a, int b, void *c, void *d, int e)
 
     if (i != -1)
     {
-        j = sub_0801DC04(c, (s16)(int)d, (s16)e);
+        j = (s8)sub_0801DC04(c, (s16)(int)d, (s16)e);
 
         if (j == -1)
             return j;
